requirements.txt
fastapi
uvicorn
SQLAlchemy
mysql-connector-python
python-dotenv
requests
urllib3
scikit-learn
pandas
requests-cache
tqdm
scikit-learn
urllib3
